% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_spatial_variability.R
\name{analyze_spatial_variability}
\alias{analyze_spatial_variability}
\title{Analyser la variabilite spatiale du SOC}
\usage{
analyze_spatial_variability(
  sol_stock,
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/data"
)
}
\arguments{
\item{sol_stock}{sf object avec colonne SOC_stock_tCha}

\item{output_dir}{dossier de sauvegarde}
}
\value{
liste avec stats, moran et sol_classified
}
\description{
Calcule statistiques descriptives et indice de Moran.
}
\examples{
# res <- analyze_spatial_variability(sol_stock)
}
