% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_soc_stock.R
\name{calculate_soc_stock}
\alias{calculate_soc_stock}
\title{Calculer le stock de carbone organique}
\usage{
calculate_soc_stock(
  sol_data,
  depth = 30,
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/data"
)
}
\arguments{
\item{sol_data}{sf object avec colonnes SOC_mean et BD_mean}

\item{depth}{profondeur en cm (defaut: 30)}

\item{output_dir}{dossier de sauvegarde}
}
\value{
sf object avec colonne SOC_stock_tCha
}
\description{
Calcule le stock SOC en tC/ha.
}
\examples{
# sol_stock <- calculate_soc_stock(sol_data, depth=30)
}
