% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_sequestration_potential.R
\name{estimate_sequestration_potential}
\alias{estimate_sequestration_potential}
\title{Estimer le potentiel de sequestration carbone}
\usage{
estimate_sequestration_potential(
  sol_stock,
  scenarios = c("cover_crop", "no_tillage", "organic_fert"),
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/data"
)
}
\arguments{
\item{sol_stock}{sf object avec colonne SOC_stock_tCha}

\item{scenarios}{vecteur de scenarios}

\item{output_dir}{dossier de sauvegarde}
}
\value{
data.frame avec gains par scenario
}
\description{
Calcule les gains potentiels en SOC par scenario.
}
\examples{
# seq_pot <- estimate_sequestration_potential(sol_stock)
}
