% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_covariates.R
\name{extract_covariates}
\alias{extract_covariates}
\title{Extraire et joindre les covariables aux données sol}
\usage{
extract_covariates(soil_data, env_data)
}
\arguments{
\item{soil_data}{Data frame. Données sol avec parcelle_id.}

\item{env_data}{Data frame. Covariables environnementales avec parcelle_id.}
}
\value{
Data frame complet pour la modélisation.
}
\description{
Joint les covariables environnementales aux données pédologiques.
}
