% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farmCarbonR-package.R
\docType{package}
\name{farmCarbonR-package}
\alias{farmCarbonR-package}
\alias{farmCarbonR}
\title{farmCarbonR: Estimation des stocks de carbone agricole}
\description{
Package pour estimer et cartographier les stocks SOC
  en contexte agricole marocain avec Random Forest.
}
\author{
\strong{Maintainer}: Ton Nom Ton Prenom \email{ton.email@example.com}

Authors:
\itemize{
  \item Ton Nom Ton Prenom \email{ton.email@example.com}
}

}
