% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Générer un rapport scientifique farmCarbonR}
\usage{
generate_report(
  data,
  rf_result = NULL,
  soc_map = NULL,
  output_dir = "outputs",
  output_format = "html",
  titre = "Rapport scientifique farmCarbonR"
)
}
\description{
Générer un rapport scientifique farmCarbonR
}
