% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_area.R
\name{get_study_area}
\alias{get_study_area}
\title{Recuperer les parametres de la zone d'etude}
\usage{
get_study_area()
}
\value{
Liste des parametres
}
\description{
Recuperer les parametres de la zone d'etude
}
