% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ndvi.R
\name{load_ndvi}
\alias{load_ndvi}
\title{Charger NDVI Maroc}
\usage{
load_ndvi(
  lon_min = -6.5,
  lon_max = -4.5,
  lat_min = 31,
  lat_max = 33,
  ndvi_value = 0.35,
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/data"
)
}
\arguments{
\item{lon_min}{longitude minimum (defaut: -6.5)}

\item{lon_max}{longitude maximum (defaut: -4.5)}

\item{lat_min}{latitude minimum  (defaut: 31.0)}

\item{lat_max}{latitude maximum  (defaut: 33.0)}

\item{ndvi_value}{valeur NDVI par defaut (defaut: 0.35)}

\item{output_dir}{dossier de sauvegarde}
}
\value{
SpatRaster avec couche NDVI
}
\description{
Cree un raster NDVI sur la zone d etude marocaine.
}
\examples{
# ndvi <- load_ndvi()
}
