% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soc_map.R
\name{plot_soc_map}
\alias{plot_soc_map}
\title{Visualiser la carte SOC}
\usage{
plot_soc_map(
  soc_map,
  sol_stock = NULL,
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/outputs",
  filename = "soc_map.png",
  title = "Stock de Carbone Organique (tC/ha)"
)
}
\arguments{
\item{soc_map}{SpatRaster retourne par predict_soc_map()}

\item{sol_stock}{sf object des parcelles (optionnel)}

\item{output_dir}{dossier de sauvegarde}

\item{filename}{nom du fichier PNG}

\item{title}{titre de la carte}
}
\value{
objet ggplot2
}
\description{
Produit une carte ggplot2 du stock SOC predit.
}
\examples{
# p <- plot_soc_map(soc_map, sol_stock)
}
