% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_farm_practices}
\alias{sample_farm_practices}
\title{Sample Farm Practices Data}
\format{
A data frame with 50 rows and 6 variables:
\describe{
  \item{parcelle_id}{Identifiant parcelle}
  \item{travail_sol}{Type de travail du sol}
  \item{couvert_vegetal}{Presence couvert vegetal}
  \item{irrigation}{Mode irrigation}
  \item{fertilisation_organique}{Type fertilisation organique}
  \item{rotation}{Type de rotation culturale}
}
}
\source{
Recensement agricole Maroc 2016 (FAO)
}
\usage{
data(sample_farm_practices)
}
\description{
Sample Farm Practices Data
}
\keyword{datasets}
