% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_soil_data}
\alias{sample_soil_data}
\title{Sample Soil Data for Moroccan Agricultural Sites}
\format{
A data frame with 50 rows and 14 variables:
\describe{
  \item{parcelle_id}{Identifiant parcelle}
  \item{lon}{Longitude (WGS84)}
  \item{lat}{Latitude (WGS84)}
  \item{SOC_mean}{Carbone organique du sol (g/kg)}
  \item{BD_mean}{Densite apparente (g/cm3)}
  \item{clay}{Argile (pct)}
  \item{sand}{Sable (pct)}
  \item{silt}{Limon (pct)}
  \item{depth}{Profondeur (cm)}
  \item{rock_fragment}{Fragments grossiers}
  \item{texture}{Classe texturale}
  \item{temp}{Temperature moyenne annuelle (degC)}
  \item{precip}{Precipitations annuelles (mm)}
  \item{alt}{Altitude (m)}
}
}
\source{
SoilGrids ISRIC + WorldClim v2.1 + SRTM
}
\usage{
data(sample_soil_data)
}
\description{
Sample Soil Data for Moroccan Agricultural Sites
}
\keyword{datasets}
