% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_area.R
\name{set_study_area}
\alias{set_study_area}
\title{Definir la zone d'etude}
\usage{
set_study_area(
  lon_min = -6.5,
  lon_max = -4.5,
  lat_min = 31,
  lat_max = 33,
  start_date = NULL,
  end_date = NULL,
  country = "MAR",
  output_dir = "data"
)
}
\arguments{
\item{lon_min}{Longitude minimale}

\item{lon_max}{Longitude maximale}

\item{lat_min}{Latitude minimale}

\item{lat_max}{Latitude maximale}

\item{start_date}{Date debut (format YYYY-MM-DD)}

\item{end_date}{Date fin (format YYYY-MM-DD)}

\item{country}{Code pays ISO 3166-1 alpha-3 (ex: "MAR")}

\item{output_dir}{Dossier de sortie}
}
\value{
Liste invisible des parametres
}
\description{
Enregistre les parametres de la zone dans l'environnement interne.
}
