% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_rf_model.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Entrainer le modele Random Forest}
\usage{
train_rf_model(
  train_data,
  test_data = NULL,
  target = "SOC_stock_tCha",
  ntree = 500,
  mtry = NULL,
  seed = 42
)
}
\arguments{
\item{train_data}{Data frame d'entrainement.}

\item{test_data}{Data frame de test (optionnel).}

\item{target}{Variable cible. Par defaut "SOC_stock_tCha".}

\item{ntree}{Nombre d'arbres. Par defaut 500.}

\item{mtry}{Variables par split. Par defaut NULL (auto).}

\item{seed}{Graine. Par defaut 42.}
}
\value{
Liste : model, importance, oob_error, rmse_train, rmse_test,
  r2_train, r2_test, predicteurs.
}
\description{
Entrainer le modele Random Forest
}
