% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_spatial_variability.R
\name{analyze_spatial_variability}
\alias{analyze_spatial_variability}
\title{Analyser la variabilité spatiale du SOC}
\usage{
analyze_spatial_variability(data, krigeage = FALSE)
}
\arguments{
\item{data}{Data frame avec SOC_stock_tCha, lon, lat.}

\item{krigeage}{Logique. Effectuer le krigeage ? Par défaut FALSE.}
}
\value{
Liste : stats, variogram_df, moran.
}
\description{
Calcule le variogramme, l'autocorrélation spatiale (Moran)
et optionnellement un krigeage simplifié.
}
