% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ndvi.R
\name{load_ndvi}
\alias{load_ndvi}
\title{Charger le NDVI depuis MODIS}
\usage{
load_ndvi(coords, start_date = "2023-01-01", end_date = "2023-12-31")
}
\arguments{
\item{coords}{Data frame avec colonnes parcelle_id, lon, lat.}

\item{start_date}{Caractère. Date début. Par défaut "2023-01-01".}

\item{end_date}{Caractère. Date fin. Par défaut "2023-12-31".}
}
\value{
Data frame avec colonnes parcelle_id, lon, lat, ndvi.
}
\description{
Télécharge le NDVI réel depuis MODIS (MOD13A3) via MODISTools.
}
