% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_importance.R
\name{plot_feature_importance}
\alias{plot_feature_importance}
\title{Visualiser l importance des variables}
\usage{
plot_feature_importance(
  rf_result,
  output_dir = "C:/Users/PC Paradise/Desktop/farmCarbonR/outputs",
  filename = "feature_importance.png"
)
}
\arguments{
\item{rf_result}{liste retournee par train_rf_model()}

\item{output_dir}{dossier de sauvegarde}

\item{filename}{nom du fichier PNG}
}
\value{
objet ggplot2
}
\description{
Graphique en barres de l importance des variables RF.
}
\examples{
# p <- plot_feature_importance(rf_result)
}
